myModel-classification/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.14.0
scikit-learn==1.3.0
pandas==2.1.1
python-dotenv==1.0.0 